// libs/store/Chat/actions-type.tsx
export enum ChatActionType {
    JOIN_CHAT_ROOM = 'chat/JOIN_CHAT_ROOM',
    JOIN_CHAT_ROOM_SUCCESS = 'chat/JOIN_CHAT_ROOM_SUCCESS',
    JOIN_CHAT_ROOM_FAILED = 'chat/JOIN_CHAT_ROOM_FAILED ',
    ADD_CHAT_MESSAGE = 'chat/ADD_CHAT_MESSAGE',
    ADD_CHAT_MESSAGE_SUCCESS = 'chat/ADD_CHAT_MESSAGE_SUCCESS',
    ADD_CHAT_MESSAGE_FAILED = 'chat/ADD_CHAT_MESSAGE_FAILED',
    GET_CHAT_MESSAGES = 'chat/GET_CHAT_MESSAGES ',
}